requirements.txt
# Core ML / numerical
numpy>=1.26
scipy>=1.11
pandas>=2.1
scikit-learn>=1.4

# LLMs / activations
torch>=2.2
transformers>=4.46
accelerate>=0.34
safetensors>=0.4
datasets>=2.20
huggingface-hub>=0.24
sentencepiece>=0.2
sae-lens>=6.28

# Experiment / config / utilities
PyYAML>=6.0
tqdm>=4.66
joblib>=1.3

# Metrics / plotting (lightweight; optional but useful for analysis)
matplotlib>=3.8
seaborn>=0.13

# Testing / quality (research reproducibility support)
pytest>=8.0
ruff>=0.6
